@@ -1,15 +1,15 @@
-module devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval
+module github.com/kweaver-ai/adp/context-loader/agent-retrieval
 
-go 1.23.7
+go 1.24.11
 
 require (
-	devops.aishu.cn/AISHUDevOps/DIP/_git/mdl-go-lib v0.0.31
-	devops.aishu.cn/AISHUDevOps/ONE-Architecture/_git/TelemetrySDK-Go.git/span/v2 v2.9.1
 	github.com/creasty/defaults v1.8.0
-	github.com/gin-gonic/gin v1.10.1
-	github.com/go-playground/validator/v10 v10.26.0
+	github.com/gin-gonic/gin v1.11.0
+	github.com/go-playground/validator/v10 v10.30.1
 	github.com/go-redis/redis/v8 v8.11.5
 	github.com/json-iterator/go v1.1.12
+	github.com/kweaver-ai/TelemetrySDK-Go/span/v2 v2.10.3
+	github.com/kweaver-ai/kweaver-go-lib v1.0.2
 	github.com/nicksnyder/go-i18n/v2 v2.6.0
 	github.com/panjf2000/ants v1.3.0
 	github.com/pkg/errors v0.9.1
