@@ -1,3 +1,8 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 // Package telemetry 可观测性相关包
 package telemetry
 
@@ -5,8 +10,8 @@ import (
 	"context"
 	"fmt"
 
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/interfaces"
-	o11y "devops.aishu.cn/AISHUDevOps/DIP/_git/mdl-go-lib/observability"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/interfaces"
+	o11y "github.com/kweaver-ai/kweaver-go-lib/observability"
 )
 
 // LogExporterType 日志导出类型
