@@ -1,81 +1,86 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 package interfaces
 
 import "context"
 
-// ChatRequest 聊天请求
+// ChatRequest chat request
 type ChatRequest struct {
 	AgentKey     string         `json:"agent_key"`                                // agent key
-	AgentVersion string         `json:"agent_version,omitempty" default:"latest"` // agent 版本,默认请求最新版本
-	Stream       bool           `json:"stream"`                                   // 是否流式返回
-	Query        string         `json:"query"`                                    // 用户提问问题
-	CustomQuerys map[string]any `json:"custom_querys"`                            // 输入配置中的自定义输入变量
+	AgentVersion string         `json:"agent_version,omitempty" default:"latest"` // agent version, default request latest version
+	Stream       bool           `json:"stream"`                                   // Whether to return stream
+	Query        string         `json:"query"`                                    // User question
+	CustomQuerys map[string]any `json:"custom_querys"`                            // Custom input variables in input configuration
 }
 
 type Answer struct {
-	Text string `json:"text"` // 回答内容
+	Text string `json:"text"` // Answer content
 }
 
-// FinalAnswer 最终结果
+// FinalAnswer Final Result
 type FinalAnswer struct {
-	Query  string  `json:"query"`  // 原始查询
-	Answer *Answer `json:"answer"` // 回答
+	Query  string  `json:"query"`  // Original Query
+	Answer *Answer `json:"answer"` // Answer
 }
 
 type ChatContent struct {
-	FinalAnswer *FinalAnswer `json:"final_answer"` // 最终结果
+	FinalAnswer *FinalAnswer `json:"final_answer"` // Final Result
 }
 
-// ChatMessage 消息内容
+// ChatMessage Message Content
 type ChatMessage struct {
-	ID             string       `json:"id"`              // 消息ID
-	ConversationID string       `json:"conversation_id"` // 会话ID
-	Role           string       `json:"role"`            // 消息角色
-	Content        *ChatContent `json:"content"`         // 消息内容
+	ID             string       `json:"id"`              // Message ID
+	ConversationID string       `json:"conversation_id"` // Conversation ID
+	Role           string       `json:"role"`            // Message Role
+	Content        *ChatContent `json:"content"`         // Message Content
 }
 
-// ChatResponse 聊天响应
+// ChatResponse Chat Response
 type ChatResponse struct {
-	ConversationID     string       `json:"conversation_id"`      // 会话ID
-	UserMessageID      string       `json:"user_message_id"`      // 用户消息ID
-	AssistantMessageID string       `json:"assistant_message_id"` // 助手消息ID
-	Message            *ChatMessage `json:"message"`              // 消息内容
+	ConversationID     string       `json:"conversation_id"`      // Conversation ID
+	UserMessageID      string       `json:"user_message_id"`      // User Message ID
+	AssistantMessageID string       `json:"assistant_message_id"` // Assistant Message ID
+	Message            *ChatMessage `json:"message"`              // Message Content
 }
 
-// ConceptIntentionAnalysisAgentReq 概念意图分析智能体请求
+// ConceptIntentionAnalysisAgentReq Concept Intention Analysis Agent Request
 type ConceptIntentionAnalysisAgentReq struct {
-	PreviousQueries []string `json:"previous_queries"` // 历史查询
-	Query           string   `json:"query"`            // 当前查询
-	KnID            string   `json:"kn_id"`            // 知识网络ID
+	PreviousQueries []string `json:"previous_queries"` // History Queries
+	Query           string   `json:"query"`            // Current Query
+	KnID            string   `json:"kn_id"`            // Knowledge Network ID
 }
 
-// ConceptRetrievalStrategistReq 概念召回策略智能体请求
+// ConceptRetrievalStrategistReq Concept Retrieval Strategist Agent Request
 type ConceptRetrievalStrategistReq struct {
-	QueryParam      *ConceptRetrievalStrategistQueryParam `json:"query_param"`      // 概念召回策略智能体Query
-	PreviousQueries []string                              `json:"previous_queries"` // 历史查询
-	KnID            string                                `json:"kn_id"`            // 知识网络ID
+	QueryParam      *ConceptRetrievalStrategistQueryParam `json:"query_param"`      // Concept Retrieval Strategist Agent Query
+	PreviousQueries []string                              `json:"previous_queries"` // History Queries
+	KnID            string                                `json:"kn_id"`            // Knowledge Network ID
 }
 
-// ConceptRetrievalStrategistQueryParam 概念召回策略智能体Query
+// ConceptRetrievalStrategistQueryParam Concept Retrieval Strategist Agent Query
 type ConceptRetrievalStrategistQueryParam struct {
-	OriginalQuery        string               `json:"original_query"`         // 原始查询
-	CurrentIntentSegment *SemanticQueryIntent `json:"current_intent_segment"` // 当前意图片段
-	ConceptCandidates    []*ConceptResult     `json:"concept_candidates"`     // 概念候选集
+	OriginalQuery        string               `json:"original_query"`         // Original Query
+	CurrentIntentSegment *SemanticQueryIntent `json:"current_intent_segment"` // Current Intent Segment
+	ConceptCandidates    []*ConceptResult     `json:"concept_candidates"`     // Concept Candidates
 }
 
-// AgentApp 智能体接口
+// AgentApp Agent Interface
 type AgentApp interface {
 	APIChat(ctx context.Context, req *ChatRequest) (*ChatResponse, error)
-	// ConceptIntentionAnalysisAgent 概念意图分析智能体
+	// ConceptIntentionAnalysisAgent Concept Intention Analysis Agent
 	ConceptIntentionAnalysisAgent(ctx context.Context, req *ConceptIntentionAnalysisAgentReq) (*QueryUnderstanding, error)
-	// ConceptRetrievalStrategistAgent 概念召回策略智能体
+	// ConceptRetrievalStrategistAgent Concept Retrieval Strategist Agent
 	ConceptRetrievalStrategistAgent(ctx context.Context, req *ConceptRetrievalStrategistReq) ([]*SemanticQueryStrategy, error)
-	// MetricDynamicParamsGeneratorAgent Metric 动态参数生成智能体
+	// MetricDynamicParamsGeneratorAgent Metric Dynamic Params Generator Agent
 	MetricDynamicParamsGeneratorAgent(ctx context.Context, req *MetricDynamicParamsGeneratorReq) (dynamicParams map[string]any, missingParams *MissingPropertyParams, err error)
-	// OperatorDynamicParamsGeneratorAgent Operator 动态参数生成智能体
+	// OperatorDynamicParamsGeneratorAgent Operator Dynamic Params Generator Agent
 	OperatorDynamicParamsGeneratorAgent(ctx context.Context, req *OperatorDynamicParamsGeneratorReq) (dynamicParams map[string]any, missingParams *MissingPropertyParams, err error)
 }
 
-// MetricDynamicParamsGeneratorReq Metric 动态参数生成请求
+// MetricDynamicParamsGeneratorReq Metric Dynamic Params Generator Request
 type MetricDynamicParamsGeneratorReq struct {
 	LogicProperty     *LogicPropertyDef `json:"logic_property"`
 	Query             string            `json:"query"`
@@ -85,12 +90,12 @@ type MetricDynamicParamsGeneratorReq struct {
 	Timezone          string            `json:"timezone,omitempty"`
 }
 
-// OperatorDynamicParamsGeneratorReq Operator 动态参数生成请求
+// OperatorDynamicParamsGeneratorReq Operator Dynamic Params Generator Request
 type OperatorDynamicParamsGeneratorReq struct {
 	OperatorId        string            `json:"operator_id"`
 	LogicProperty     *LogicPropertyDef `json:"logic_property"`
 	Query             string            `json:"query"`
 	UniqueIdentities  []map[string]any  `json:"unique_identities"`
 	AdditionalContext string            `json:"additional_context,omitempty"`
-	// ObjectInstances 已移除，对象实例信息通过 AdditionalContext 传递
+	// ObjectInstances removed, object instance information is passed via AdditionalContext
 }
