@@ -1,32 +1,37 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 package interfaces
 
 import "context"
 
-// 业务知识网络概念类型
+// KnConceptType Knowledge Network Concept Type
 type KnConceptType string
 
 const (
-	KnConceptTypeObject   KnConceptType = "object_type"   // 对象类
-	KnConceptTypeRelation KnConceptType = "relation_type" // 关系类
-	KnConceptTypeAction   KnConceptType = "action_type"   // 行动类
+	KnConceptTypeObject   KnConceptType = "object_type"   // Object Type
+	KnConceptTypeRelation KnConceptType = "relation_type" // Relation Type
+	KnConceptTypeAction   KnConceptType = "action_type"   // Action Type
 )
 
-// QueryObjectInstancesReq 检索对象详细数据请求对象
+// QueryObjectInstancesReq Request object for querying detailed object instances
 type QueryObjectInstancesReq struct {
-	KnID               string       `form:"kn_id"`                          // 知识网络ID
-	OtID               string       `form:"ot_id"`                          // 对象类ID
-	IncludeTypeInfo    bool         `form:"include_type_info"`              //是否包含对象类信息
-	IncludeLogicParams bool         `form:"include_logic_params"`           // 包含逻辑属性的计算参数，默认false，不包含。
-	Cond               *KnCondition `json:"condition"`                      // 检索条件
-	Limit              int          `json:"limit" validate:"min=1,max=100"` //返回的数量，默认值 10。范围 1-100
+	KnID               string       `form:"kn_id"`                          // Knowledge Network ID
+	OtID               string       `form:"ot_id"`                          // Object Type ID
+	IncludeTypeInfo    bool         `form:"include_type_info"`              // Whether to include object type info
+	IncludeLogicParams bool         `form:"include_logic_params"`           // Include calculation parameters for logic properties, default false
+	Cond               *KnCondition `json:"condition"`                      // Retrieval conditions
+	Limit              int          `json:"limit" validate:"min=1,max=100"` // Quantity limit, default 10, range 1-100
 }
 
 type QueryObjectInstancesResp struct {
-	Data          []any          `json:"datas"`       // 对象实例列表
-	ObjectConcept map[string]any `json:"object_type"` // 对象类型
+	Data          []any          `json:"datas"`       // List of object instances
+	ObjectConcept map[string]any `json:"object_type"` // Object type definition
 }
 
-// QueryLogicPropertiesReq 查询逻辑属性值请求
+// QueryLogicPropertiesReq Request for querying logic properties values
 type QueryLogicPropertiesReq struct {
 	KnID             string                   `json:"kn_id"`
 	OtID             string                   `json:"ot_id"`
@@ -35,39 +40,39 @@ type QueryLogicPropertiesReq struct {
 	DynamicParams    map[string]interface{}   `json:"dynamic_params"`
 }
 
-// QueryLogicPropertiesResp 查询逻辑属性值响应
+// QueryLogicPropertiesResp Response for querying logic properties values
 type QueryLogicPropertiesResp struct {
 	Datas []map[string]interface{} `json:"datas"`
 }
 
-// QueryInstanceSubgraphReq 子图查询请求
+// QueryInstanceSubgraphReq Subgraph query request
 type QueryInstanceSubgraphReq struct {
-	// Path 参数
+	// Path parameters
 	KnID string `form:"kn_id"`
 
-	// Query 参数
+	// Query parameters
 	IncludeLogicParams bool `form:"include_logic_params"`
 
-	// Body 参数 - 使用 interface{} 避免明确定义结构
-	// 对应 ontology-query 接口的 SubGraphQueryBaseOnTypePath 结构
+	// Body parameters - use interface{} to avoid explicit struct definition
+	// Corresponds to SubGraphQueryBaseOnTypePath struct in ontology-query interface
 	RelationTypePaths interface{} `json:"relation_type_paths"`
 }
 
-// QueryInstanceSubgraphResp 子图查询响应
+// QueryInstanceSubgraphResp Subgraph query response
 type QueryInstanceSubgraphResp struct {
-	// 使用 interface{} 直接返回底层接口的原始结构
-	// 对应 ontology-query 接口的 PathEntries 结构
+	// Use interface{} to directly return the original structure from the underlying interface
+	// Corresponds to PathEntries struct in ontology-query interface
 	Entries interface{} `json:"entries"`
 }
 
-// DrivenOntologyQuery 本体查询接口
+// DrivenOntologyQuery Ontology query interface
 type DrivenOntologyQuery interface {
-	// QueryObjectInstances 检索指定对象类的对象的详细数据
+	// QueryObjectInstances retrieves detailed data of objects for a specified object class
 	QueryObjectInstances(ctx context.Context, req *QueryObjectInstancesReq) (resp *QueryObjectInstancesResp, err error)
-	// QueryLogicProperties 查询逻辑属性值
+	// QueryLogicProperties queries logic property values
 	QueryLogicProperties(ctx context.Context, req *QueryLogicPropertiesReq) (resp *QueryLogicPropertiesResp, err error)
-	// QueryActions 查询行动
+	// QueryActions queries actions
 	QueryActions(ctx context.Context, req *QueryActionsRequest) (resp *QueryActionsResponse, err error)
-	// QueryInstanceSubgraph 查询对象子图
+	// QueryInstanceSubgraph queries object subgraph
 	QueryInstanceSubgraph(ctx context.Context, req *QueryInstanceSubgraphReq) (resp *QueryInstanceSubgraphResp, err error)
 }
