@@ -1,85 +1,90 @@
-// Package interfaces 定义业务知识网络行动召回接口
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
+// Package interfaces defines interfaces for business knowledge network action recall
 package interfaces
 
 import "context"
 
-// ==================== 常量定义 ====================
+// ==================== Constant Definitions ====================
 
 const (
-	// ResultProcessStrategyKnActionRecall 业务知识网络行动召回结果处理策略
+	// ResultProcessStrategyKnActionRecall Result processing strategy for knowledge network action recall
 	ResultProcessStrategyKnActionRecall = "kn_action_recall"
 )
 
-// ActionSource 类型常量
+// ActionSource Type Constants
 const (
-	// ActionSourceTypeTool 工具类型的行动源
+	// ActionSourceTypeTool Tool type action source
 	ActionSourceTypeTool = "tool"
-	// ActionSourceTypeMCP MCP 类型的行动源（下个版本支持）
+	// ActionSourceTypeMCP MCP type action source (supported in next version)
 	ActionSourceTypeMCP = "mcp"
 )
 
-// ==================== 请求和响应结构 ====================
+// ==================== Request and Response Structures ====================
 
-// KnActionRecallRequest 业务知识网络行动召回请求
+// KnActionRecallRequest Knowledge Network Action Recall Request
 type KnActionRecallRequest struct {
 	// Query Parameters
-	KnID string `json:"kn_id" validate:"required"` // 业务知识网络ID
-	AtID string `json:"at_id" validate:"required"` // 行动类型ID
+	KnID string `json:"kn_id" validate:"required"` // Knowledge Network ID
+	AtID string `json:"at_id" validate:"required"` // Action Type ID
 
 	// Request Body
-	UniqueIdentity map[string]interface{} `json:"unique_identity" validate:"required,min=1"` // 对象唯一标识
+	UniqueIdentity map[string]interface{} `json:"unique_identity" validate:"required,min=1"` // Object Unique Identity
 
-	// Header 字段
+	// Header Fields
 	AccountID   string `json:"-" header:"x-account-id" validate:"required"`
 	AccountType string `json:"-" header:"x-account-type" validate:"required"`
 }
 
-// KnActionRecallResponse 业务知识网络行动召回响应
+// KnActionRecallResponse Knowledge Network Action Recall Response
 type KnActionRecallResponse struct {
-	Headers      map[string]string `json:"headers"` // HTTP Header 参数
+	Headers      map[string]string `json:"headers"` // HTTP Header Parameters
 	DynamicTools []KnDynamicTool   `json:"_dynamic_tools"`
 }
 
-// KnDynamicTool 动态工具定义
+// KnDynamicTool Dynamic Tool Definition
 type KnDynamicTool struct {
-	Name            string                 `json:"name"`              // 工具名称
-	Description     string                 `json:"description"`       // 工具描述
+	Name            string                 `json:"name"`              // Tool Name
+	Description     string                 `json:"description"`       // Tool Description
 	Parameters      map[string]interface{} `json:"parameters"`        // OpenAI Function Call Schema
-	ApiURL          string                 `json:"api_url"`           // 工具执行代理URL
-	OriginalSchema  map[string]interface{} `json:"original_schema"`   // 原始 OpenAPI 定义
-	FixedParams     KnFixedParams          `json:"fixed_params"`      // 固定参数
-	ApiCallStrategy string                 `json:"api_call_strategy"` // 结果处理策略， 固定值为 kn_action_recall
+	ApiURL          string                 `json:"api_url"`           // Tool Execution Proxy URL
+	OriginalSchema  map[string]interface{} `json:"original_schema"`   // Original OpenAPI Definition
+	FixedParams     KnFixedParams          `json:"fixed_params"`      // Fixed Parameters
+	ApiCallStrategy string                 `json:"api_call_strategy"` // Result Processing Strategy, fixed value: kn_action_recall
 }
 
-// KnFixedParams 固定参数结构
+// KnFixedParams Fixed Parameters Structure
 type KnFixedParams struct {
-	Header map[string]interface{} `json:"header"` // HTTP Header 参数
-	Path   map[string]interface{} `json:"path"`   // URL Path 参数
-	Query  map[string]interface{} `json:"query"`  // URL Query 参数
-	Body   map[string]interface{} `json:"body"`   // Request Body 参数
+	Header map[string]interface{} `json:"header"` // HTTP Header Parameters
+	Path   map[string]interface{} `json:"path"`   // URL Path Parameters
+	Query  map[string]interface{} `json:"query"`  // URL Query Parameters
+	Body   map[string]interface{} `json:"body"`   // Request Body Parameters
 }
 
-// ==================== 行动查询相关结构 ====================
+// ==================== Action Query Related Structures ====================
 
-// QueryActionsRequest 行动查询请求
+// QueryActionsRequest Action Query Request
 type QueryActionsRequest struct {
 	KnID                string                   `json:"kn_id"`
 	AtID                string                   `json:"at_id"`
 	UniqueIdentities    []map[string]interface{} `json:"unique_identities"`
 	IncludeTypeInfo     bool                     `json:"include_type_info"`
-	XHTTPMethodOverride string                   `json:"-"` // 固定为 GET
+	XHTTPMethodOverride string                   `json:"-"` // Fixed to GET
 }
 
-// QueryActionsResponse 行动查询响应
+// QueryActionsResponse Action Query Response
 type QueryActionsResponse struct {
-	ActionType   *ActionTypeInfo `json:"action_type,omitempty"` // 行动类信息
-	ActionSource *ActionSource   `json:"action_source"`         // 行动源
-	Actions      []ActionParams  `json:"actions"`               // 行动参数列表
+	ActionType   *ActionTypeInfo `json:"action_type,omitempty"` // Action Type Info
+	ActionSource *ActionSource   `json:"action_source"`         // Action Source
+	Actions      []ActionParams  `json:"actions"`               // Action Parameters List
 	TotalCount   int             `json:"total_count"`
 	OverallMs    int             `json:"overall_ms"`
 }
 
-// ActionTypeInfo 行动类信息
+// ActionTypeInfo Action Type Info
 type ActionTypeInfo struct {
 	ID           string                 `json:"id"`
 	Name         string                 `json:"name"`
@@ -91,7 +96,7 @@ type ActionTypeInfo struct {
 	Schedule     map[string]interface{} `json:"schedule"`
 }
 
-// ActionTypeParam 行动类型参数
+// ActionTypeParam Action Type Parameter
 type ActionTypeParam struct {
 	Name      string `json:"name"`
 	Type      string `json:"type"`
@@ -100,23 +105,23 @@ type ActionTypeParam struct {
 	Value     string `json:"value,omitempty"`
 }
 
-// ActionSource 行动源
+// ActionSource Action Source
 type ActionSource struct {
 	Type   string `json:"type"`    // tool/mcp
-	BoxID  string `json:"box_id"`  // 工具箱ID
-	ToolID string `json:"tool_id"` // 工具ID
+	BoxID  string `json:"box_id"`  // Tool Box ID
+	ToolID string `json:"tool_id"` // Tool ID
 }
 
-// ActionParams 行动参数
+// ActionParams Action Parameters
 type ActionParams struct {
-	Parameters    map[string]interface{} `json:"parameters"`     // 已实例化的参数
-	DynamicParams map[string]interface{} `json:"dynamic_params"` // 动态参数（值为null）
+	Parameters    map[string]interface{} `json:"parameters"`     // Instantiated Parameters
+	DynamicParams map[string]interface{} `json:"dynamic_params"` // Dynamic Parameters (value is null)
 }
 
-// ==================== Service 接口 ====================
+// ==================== Service Interfaces ====================
 
-// IKnActionRecallService 业务知识网络行动召回服务接口
+// IKnActionRecallService Knowledge Network Action Recall Service Interface
 type IKnActionRecallService interface {
-	// GetActionInfo 获取行动信息（行动召回）
+	// GetActionInfo gets action information (action recall)
 	GetActionInfo(ctx context.Context, req *KnActionRecallRequest) (*KnActionRecallResponse, error)
 }
