@@ -1,120 +1,125 @@
-// Package interfaces 定义kn-logic-property-resolver接口
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
+// Package interfaces defines kn-logic-property-resolver interfaces
 package interfaces
 
 import "context"
 
-// ==================== 请求和响应结构 ====================
+// ==================== Request and Response Structures ====================
 
-// ResolveLogicPropertiesRequest 逻辑属性解析请求
+// ResolveLogicPropertiesRequest Logic property resolution request
 type ResolveLogicPropertiesRequest struct {
-	// 必填字段
+	// Required fields
 	KnID             string                   `json:"kn_id" validate:"required"`
 	OtID             string                   `json:"ot_id" validate:"required"`
 	Query            string                   `json:"query"`
 	UniqueIdentities []map[string]interface{} `json:"unique_identities" validate:"required"`
 	Properties       []string                 `json:"properties" validate:"required"`
 
-	// 可选字段
+	// Optional fields
 	AdditionalContext string          `json:"additional_context,omitempty"`
 	Options           *ResolveOptions `json:"options,omitempty"`
 
-	// Header 字段
+	// Header fields
 	AccountID   string `json:"-" header:"x-account-id"`
 	AccountType string `json:"-" header:"x-account-type"`
 }
 
-// ResolveOptions 可选配置
+// ResolveOptions Optional configuration
 type ResolveOptions struct {
 	ReturnDebug     bool `json:"return_debug" default:"false"`
 	MaxRepairRounds int  `json:"max_repair_rounds" default:"1"`
 	MaxConcurrency  int  `json:"max_concurrency" default:"4"`
 }
 
-// ResolveLogicPropertiesResponse 逻辑属性解析响应
+// ResolveLogicPropertiesResponse Logic property resolution response
 type ResolveLogicPropertiesResponse struct {
 	Datas []map[string]any `json:"datas"`
 
-	// Debug 信息（仅在 return_debug=true 时返回）
+	// Debug information (returned only when return_debug=true)
 	Debug *ResolveDebugInfo `json:"debug,omitempty"`
 }
 
-// ResolveDebugInfo Debug 信息（改进版）
+// ResolveDebugInfo Debug information (improved)
 type ResolveDebugInfo struct {
-	// Agent 生成的所有动态参数
+	// All dynamic parameters generated by Agent
 	DynamicParams map[string]any `json:"dynamic_params,omitempty"`
 
-	// Agent 调用信息（按 property 分组）
+	// Agent call info (grouped by property)
 	AgentInfo map[string]*AgentInfo `json:"agent_info,omitempty"`
 
-	// 服务器时间戳
+	// Server timestamp
 	NowMs int64 `json:"now_ms,omitempty"`
 
-	// 警告信息
+	// Warning messages
 	Warnings []string `json:"warnings,omitempty"`
 
-	// 追踪 ID
+	// Trace ID
 	TraceID string `json:"trace_id,omitempty"`
 }
 
-// AgentInfo Agent 调用信息
+// AgentInfo Agent call information
 type AgentInfo struct {
-	// 属性类型
-	PropertyType string `json:"property_type"` // "metric" 或 "operator"
+	// Property type
+	PropertyType string `json:"property_type"` // "metric" or "operator"
 
-	// Agent 请求参数（直接存储 Agent 请求结构）
+	// Agent request parameters (stores Agent request structure directly)
 	Request AgentRequestDebugInfo `json:"request,omitempty"`
 
-	// Agent 响应信息
+	// Agent response info
 	Response *AgentResponseDebugInfo `json:"response,omitempty"`
 }
 
-// AgentRequestDebugInfo Agent 请求调试信息
-// 直接存储 Agent 的请求结构：MetricDynamicParamsGeneratorReq 或 OperatorDynamicParamsGeneratorReq
+// AgentRequestDebugInfo Agent request debug info
+// Directly stores Agent request structure: MetricDynamicParamsGeneratorReq or OperatorDynamicParamsGeneratorReq
 type AgentRequestDebugInfo any
 
-// AgentResponseDebugInfo Agent 响应调试信息
-// 直接存储 Agent 的响应信息
+// AgentResponseDebugInfo Agent response debug info
+// Directly stores Agent response info
 type AgentResponseDebugInfo struct {
-	// Agent 成功响应：动态参数
+	// Agent success response: dynamic parameters
 	DynamicParams map[string]any `json:"dynamic_params,omitempty"`
 
-	// Agent 失败响应：错误信息（对应 _error 字段）
+	// Agent failure response: error message (corresponds to _error field)
 	Error string `json:"_error,omitempty"`
 }
 
-// MissingParamsError 缺参错误响应（改进版）
+// MissingParamsError Missing parameters error response (improved)
 type MissingParamsError struct {
 	ErrorCode string `json:"error_code"`
 	Message   string `json:"message"`
 
-	// 直接返回Agent生成的原始错误消息
+	// Directly return the raw error message generated by Agent
 	ErrorMsg string `json:"error_msg,omitempty"`
 
-	// Debug信息（仅在开启debug时返回）
+	// Debug info (returned only when debug is enabled)
 	Debug *ResolveDebugInfo `json:"debug,omitempty"`
 
 	Missing []MissingPropertyParams `json:"missing"`
 	TraceID string                  `json:"trace_id"`
 }
 
-// MissingPropertyParams 缺参信息（简化版）
+// MissingPropertyParams Missing parameter info (simplified)
 type MissingPropertyParams struct {
 	Property string `json:"property"`
-	// 直接返回 Agent 生成的错误消息，不再解析具体参数信息
+	// Directly return the error message generated by Agent, no longer parsing specific parameter info
 	ErrorMsg string `json:"error_msg,omitempty"`
 }
 
-// ==================== Service 接口 ====================
+// ==================== Service Interface ====================
 
-// IKnLogicPropertyResolverService 逻辑属性解析服务接口
+// IKnLogicPropertyResolverService Logic property resolver service interface
 type IKnLogicPropertyResolverService interface {
-	// ResolveLogicProperties 解析逻辑属性
+	// ResolveLogicProperties Resolve logic properties
 	ResolveLogicProperties(ctx context.Context, req *ResolveLogicPropertiesRequest) (*ResolveLogicPropertiesResponse, error)
 }
 
-// ==================== LLM 参数生成相关结构 ====================
+// ==================== LLM Parameter Generation Related Structures ====================
 
-// LLMPromptInput LLM 提示词输入
+// LLMPromptInput LLM prompt input
 type LLMPromptInput struct {
 	LogicProperty     *LogicPropertyDef `json:"logic_property"`
 	Query             string            `json:"query"`
@@ -124,11 +129,11 @@ type LLMPromptInput struct {
 	Timezone          string            `json:"timezone,omitempty"`
 }
 
-// LLMPromptOutput LLM 提示词输出
+// LLMPromptOutput LLM prompt output
 type LLMPromptOutput struct {
-	// 成功输出：{ "property_name": { "param1": value1, ... } }
+	// Success output: { "property_name": { "param1": value1, ... } }
 	DynamicParams map[string]any `json:"-"`
 
-	// 缺参输出：{ "_error": "..." }
+	// Missing parameter output: { "_error": "..." }
 	Error string `json:"_error,omitempty"`
 }
