@@ -1,44 +1,49 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 package main
 
 import (
 	"fmt"
 
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/driveradapters"
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/infra/common"
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/infra/config"
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/infra/telemetry"
-	"devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/interfaces"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/driveradapters"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/infra/common"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/infra/config"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/infra/telemetry"
+	"github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/interfaces"
 
 	"github.com/gin-gonic/gin"
 )
 
-// Server 服务
+// Server Service
 type Server struct {
-	// 健康检查
+	// Health check
 	httpHealthHandler  interfaces.HTTPRouterInterface
 	restPublicHandler  interfaces.HTTPRouterInterface
 	restPrivateHandler interfaces.HTTPRouterInterface
 	config             *config.Config
 }
 
-// Start 开启服务
+// Start starts the server
 func (s *Server) Start() {
 	gin.SetMode(gin.ReleaseMode)
 
 	go func() {
-		// 注册路由 - 健康检查
+		// Register router - health check
 		engine := gin.New()
 		engine.Use(gin.Recovery())
 		engine.UseRawPath = true
 		routerHealth := engine.Group("/health")
 		s.httpHealthHandler.RegisterRouter(routerHealth)
 
-		// 注册内部接口路由 - 算子相关接口
+		// Register internal interface router - operator related interfaces
 		routerInternalGroup := engine.Group("/api/agent-retrieval/in/v1")
 		routerInternalGroup.Use(gin.Recovery())
 		s.restPrivateHandler.RegisterRouter(routerInternalGroup)
 
-		// 注册外部路由
+		// Register external router
 		routerGroup := engine.Group("/api/agent-retrieval/v1")
 		routerGroup.Use(gin.Recovery())
 		s.restPublicHandler.RegisterRouter(routerGroup)
@@ -52,9 +57,9 @@ func (s *Server) Start() {
 }
 
 func main() {
-	// 初始化全局配置
+	// Initialize global configuration
 	config := config.NewConfigLoader()
-	// 设置错误码语言
+	// Set error code language
 	common.SetLang(config.Project.Language)
 	s := &Server{
 		config:             config,
@@ -63,7 +68,7 @@ func main() {
 		restPrivateHandler: driveradapters.NewRestPrivateHandler(config.Logger),
 	}
 	s.config.Logger.Info("start agent-retrieval server")
-	// 检查是否开启了可观测性
+	// Check if observability is enabled
 	if config.Observability.TraceEnabled {
 		defer telemetry.StopTrace()
 	}
