@@ -1,3 +1,8 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 // Code generated by MockGen. DO NOT EDIT.
 // Source: drivenadapters.go
 //
@@ -13,7 +18,7 @@ import (
 	context "context"
 	reflect "reflect"
 
-	interfaces "devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/interfaces"
+	interfaces "github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/interfaces"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -80,6 +85,41 @@ func (m *MockUserManagement) EXPECT() *MockUserManagementMockRecorder {
 	return m.recorder
 }
 
+// GetAppInfo mocks base method.
+func (m *MockUserManagement) GetAppInfo(ctx context.Context, appID string) (*interfaces.AppInfo, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetAppInfo", ctx, appID)
+	ret0, _ := ret[0].(*interfaces.AppInfo)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetAppInfo indicates an expected call of GetAppInfo.
+func (mr *MockUserManagementMockRecorder) GetAppInfo(ctx, appID any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetAppInfo", reflect.TypeOf((*MockUserManagement)(nil).GetAppInfo), ctx, appID)
+}
+
+// GetUserInfo mocks base method.
+func (m *MockUserManagement) GetUserInfo(ctx context.Context, userID string, fields ...string) (*interfaces.UserInfo, error) {
+	m.ctrl.T.Helper()
+	varargs := []any{ctx, userID}
+	for _, a := range fields {
+		varargs = append(varargs, a)
+	}
+	ret := m.ctrl.Call(m, "GetUserInfo", varargs...)
+	ret0, _ := ret[0].(*interfaces.UserInfo)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetUserInfo indicates an expected call of GetUserInfo.
+func (mr *MockUserManagementMockRecorder) GetUserInfo(ctx, userID any, fields ...any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	varargs := append([]any{ctx, userID}, fields...)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUserInfo", reflect.TypeOf((*MockUserManagement)(nil).GetUserInfo), varargs...)
+}
+
 // GetUsersInfo mocks base method.
 func (m *MockUserManagement) GetUsersInfo(ctx context.Context, userIDs, fields []string) ([]*interfaces.UserInfo, error) {
 	m.ctrl.T.Helper()
@@ -93,4 +133,73 @@ func (m *MockUserManagement) GetUsersInfo(ctx context.Context, userIDs, fields [
 func (mr *MockUserManagementMockRecorder) GetUsersInfo(ctx, userIDs, fields any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUsersInfo", reflect.TypeOf((*MockUserManagement)(nil).GetUsersInfo), ctx, userIDs, fields)
+}
+
+// GetUsersName mocks base method.
+func (m *MockUserManagement) GetUsersName(ctx context.Context, userIDs []string) (map[string]string, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "GetUsersName", ctx, userIDs)
+	ret0, _ := ret[0].(map[string]string)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// GetUsersName indicates an expected call of GetUsersName.
+func (mr *MockUserManagementMockRecorder) GetUsersName(ctx, userIDs any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "GetUsersName", reflect.TypeOf((*MockUserManagement)(nil).GetUsersName), ctx, userIDs)
+}
+
+// MockDataRetrieval is a mock of DataRetrieval interface.
+type MockDataRetrieval struct {
+	ctrl     *gomock.Controller
+	recorder *MockDataRetrievalMockRecorder
+	isgomock struct{}
+}
+
+// MockDataRetrievalMockRecorder is the mock recorder for MockDataRetrieval.
+type MockDataRetrievalMockRecorder struct {
+	mock *MockDataRetrieval
+}
+
+// NewMockDataRetrieval creates a new mock instance.
+func NewMockDataRetrieval(ctrl *gomock.Controller) *MockDataRetrieval {
+	mock := &MockDataRetrieval{ctrl: ctrl}
+	mock.recorder = &MockDataRetrievalMockRecorder{mock}
+	return mock
+}
+
+// EXPECT returns an object that allows the caller to indicate expected use.
+func (m *MockDataRetrieval) EXPECT() *MockDataRetrievalMockRecorder {
+	return m.recorder
+}
+
+// KnSearch mocks base method.
+func (m *MockDataRetrieval) KnSearch(ctx context.Context, req *interfaces.KnSearchReq) (*interfaces.KnSearchResp, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "KnSearch", ctx, req)
+	ret0, _ := ret[0].(*interfaces.KnSearchResp)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// KnSearch indicates an expected call of KnSearch.
+func (mr *MockDataRetrievalMockRecorder) KnSearch(ctx, req any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "KnSearch", reflect.TypeOf((*MockDataRetrieval)(nil).KnSearch), ctx, req)
+}
+
+// KnowledgeRerank mocks base method.
+func (m *MockDataRetrieval) KnowledgeRerank(ctx context.Context, req *interfaces.KnowledgeRerankReq) ([]*interfaces.ConceptResult, error) {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "KnowledgeRerank", ctx, req)
+	ret0, _ := ret[0].([]*interfaces.ConceptResult)
+	ret1, _ := ret[1].(error)
+	return ret0, ret1
+}
+
+// KnowledgeRerank indicates an expected call of KnowledgeRerank.
+func (mr *MockDataRetrievalMockRecorder) KnowledgeRerank(ctx, req any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "KnowledgeRerank", reflect.TypeOf((*MockDataRetrieval)(nil).KnowledgeRerank), ctx, req)
 }
