@@ -1,3 +1,8 @@
+// Copyright The kweaver.ai Authors.
+//
+// Licensed under the Apache License, Version 2.0.
+// See the LICENSE file in the project root for details.
+
 // Code generated by MockGen. DO NOT EDIT.
 // Source: infra.go
 //
@@ -14,7 +19,7 @@ import (
 	url "net/url"
 	reflect "reflect"
 
-	interfaces "devops.aishu.cn/AISHUDevOps/DIP/_git/agent-retrieval/server/interfaces"
+	interfaces "github.com/kweaver-ai/adp/context-loader/agent-retrieval/server/interfaces"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -139,20 +144,6 @@ func (mr *MockLoggerMockRecorder) Infof(format any, v ...any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	varargs := append([]any{format}, v...)
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Infof", reflect.TypeOf((*MockLogger)(nil).Infof), varargs...)
-}
-
-// Output mocks base method.
-func (m *MockLogger) Output(calldepth int, s string) error {
-	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "Output", calldepth, s)
-	ret0, _ := ret[0].(error)
-	return ret0
-}
-
-// Output indicates an expected call of Output.
-func (mr *MockLoggerMockRecorder) Output(calldepth, s any) *gomock.Call {
-	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "Output", reflect.TypeOf((*MockLogger)(nil).Output), calldepth, s)
 }
 
 // Warn mocks base method.
